refactor(playground): split accessibility utilities
