+++ packages/agent-playground/src/components/ui/accessibility/focus-visible.ts
@@ -1,0 +1,48 @@
+'use client';
+
+import { useEffect } from 'react';
+
+export function useFocusVisible() {
+  useEffect(() => {
+    let hadKeyboardEvent = true;
+
+    const handlePointerDown = () => {
+      hadKeyboardEvent = false;
+    };
+
+    const handleKeyDown = (e: KeyboardEvent) => {
+      if (e.metaKey || e.altKey || e.ctrlKey) {
+        return;
+      }
+      hadKeyboardEvent = true;
+    };
+
+    const handleFocus = (e: FocusEvent) => {
+      const target = e.target as HTMLElement;
+      if (target.matches(':focus-visible') || hadKeyboardEvent) {
+        target.classList.add('focus-visible');
+      }
+    };
+
+    const handleBlur = (e: FocusEvent) => {
+      const target = e.target as HTMLElement;
+      target.classList.remove('focus-visible');
+    };
+
+    document.addEventListener('keydown', handleKeyDown, true);
+    document.addEventListener('mousedown', handlePointerDown, true);
+    document.addEventListener('pointerdown', handlePointerDown, true);
+    document.addEventListener('touchstart', handlePointerDown, true);
+    document.addEventListener('focus', handleFocus, true);
+    document.addEventListener('blur', handleBlur, true);
+
+    return () => {
+      document.removeEventListener('keydown', handleKeyDown, true);
+      document.removeEventListener('mousedown', handlePointerDown, true);
+      document.removeEventListener('pointerdown', handlePointerDown, true);
+      document.removeEventListener('touchstart', handlePointerDown, true);
+      document.removeEventListener('focus', handleFocus, true);
+      document.removeEventListener('blur', handleBlur, true);
+    };
+  }, []);
+}
