+++ packages/agent-playground/src/components/ui/accessibility/media-preference-hooks.ts
@@ -1,0 +1,47 @@
+'use client';
+
+import { useEffect } from 'react';
+
+export function useHighContrastMode() {
+  useEffect(() => {
+    const mediaQuery = window.matchMedia('(prefers-contrast: high)');
+
+    const handleChange = (e: MediaQueryListEvent) => {
+      if (e.matches) {
+        document.documentElement.classList.add('high-contrast');
+      } else {
+        document.documentElement.classList.remove('high-contrast');
+      }
+    };
+
+    // Set initial state
+    if (mediaQuery.matches) {
+      document.documentElement.classList.add('high-contrast');
+    }
+
+    mediaQuery.addEventListener('change', handleChange);
+    return () => mediaQuery.removeEventListener('change', handleChange);
+  }, []);
+}
+
+export function useReducedMotion() {
+  useEffect(() => {
+    const mediaQuery = window.matchMedia('(prefers-reduced-motion: reduce)');
+
+    const handleChange = (e: MediaQueryListEvent) => {
+      if (e.matches) {
+        document.documentElement.classList.add('reduce-motion');
+      } else {
+        document.documentElement.classList.remove('reduce-motion');
+      }
+    };
+
+    // Set initial state
+    if (mediaQuery.matches) {
+      document.documentElement.classList.add('reduce-motion');
+    }
+
+    mediaQuery.addEventListener('change', handleChange);
+    return () => mediaQuery.removeEventListener('change', handleChange);
+  }, []);
+}
